Adding libraries and dependencies to the project.
Added Tile Manager, Math lib and Entity lib. Also Edited the cmake lists to include a "copy_resources" dependency that will be used to refresh the res folder (so that visual studio debug mode will "detect" changes in the res folder). Also added stub implementation for a simple maze level that will be used to further test the View functionality of SFML