Added basic movement with Box2D
Code still needs to be properly cleaned. Need to define a much cleaner way of handling physics in the game.
This still works though, the code baiscally applies an impulse based on the pressed direction. The problem is that the body will keep moving forward without any constraints.
One thing that should be done is make it so that it will only move forward when key is pressed and then when key is released slowly stop.