Added world textures
Both bitmap and PNG versions are available. All are designed to look fine when rotated by right-angles and placed adjacent to instances of themselves, except for the "asphalt_line" textures for which the orientation is important. You can pick randomly from the asphalt_line textures to get ones that, like the rest, look fine when placed adjacent to similar instances.